requirements.txt
# Simulation engine
numpy>=1.26
scipy>=1.12
simpy>=4.1

# API layer
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6

# LLM (Phase 3)
langchain-core>=0.2
langchain-openai>=0.1
langchain-anthropic>=0.1
langchain-community>=0.2
langchain-groq>=0.1
langchain-ollama>=0.1
python-dotenv>=1.0

# Streamlit (Phase 4)
streamlit>=1.33
httpx>=0.27

# Testing
pytest>=8.1
pytest-asyncio>=0.23
httpx>=0.27        # async test client for FastAPI
